requirements.txt
fastapi
pydantic
pydantic-settings
pydantic_core
pymongo
python-dotenv
uvicorn
starlette
fastapi-versioning
beanie
pytest
httpx
asgi_lifespan
pytest-asyncio
black
mongomock
mongomock-motor
pyjwt
python-multipart
requests

itsdangerous
Werkzeug